requirements.txt
pydirectinput
numpy
mediapipe==0.10.21
opencv-python
PyAutoGUI
pillow
joblib
scikit-learn==1.7.2
pytest
pyinstaller
coverage
sphinx
furo
pygrabber; platform_system == "Windows"
pyobjc-framework-AVFoundation; platform_system == "Darwin"
pyobjc-framework-ApplicationServices; platform_system == "Darwin"
pynput